fixed init log

+++ core/init.c
@@ -268,7 +268,7 @@ void uwsgi_setup_workers() {
 
 	total_memory *= (uwsgi.numproc + uwsgi.master_process);
 	if (uwsgi.numproc > 0)
-		uwsgi_log("mapped %lu bytes (%lu KB) for %d cores\n", total_memory, total_memory / 1024, uwsgi.cores * uwsgi.numproc);
+		uwsgi_log("mapped %llu bytes (%llu KB) for %d cores\n", (unsigned long long) total_memory, (unsigned long long) (total_memory / 1024), uwsgi.cores * uwsgi.numproc);
 
 }
